Fix unban. Switch to floats.  Theres plenty of significant values, so it uses less memory bandwidth.

+++ OverlappingModel.cs
@@ -122,7 +122,7 @@ class OverlappingModel : Model
 		T = weights.Count;
 		this.ground = (ground + T) % T;
 		patterns = new byte[T][];
-		base.weights = new double[T];
+		base.weights = new float[T];
 
 		int counter = 0;
 		foreach (long w in ordering)
